Improve timestamp parsing
The parsed dates are usually in D.M.Y format, but seems that some are
even in D.M. format.

Handle the D.M. formatted yearless dates by adding the current year to
the end to make it also D.M.Y.

Make sure that the D.M.Y formatted dates are parsed as such by passing
True to the "dayfirst" option in the call to dateutil.parser.parse.  Do
this only when there are at least two dots in the string so that M/D/Y
formatted dates should also work.