Parse message replies
The replies to a message were previously just catenated to the end of
the message.  Improve this so that the replies have their own class
ReplyMessage and they are collected to a list member named "replies" in
the Message class.  Also parse timestamp and sender from the reply
header line.

Move the message formatting to a separate mixin called _MessageWithBody
so that it can be reused for ReplyMessage too.