Fix message timestamp for messages with replies
If a message has replies, the timestamp in the message list page is not
the timestamp of the message, but the timestamp of the last reply.  Take
this into account in our code by renaming the "timestamp" field in the
MessageInfo class to "last_timestamp" and add a new "timestamp" field to
Message class.  Parse the correct timestamp from the body of the message
page.

Improve the type stubs of bs4.element.Tag with a few properties, since
the code needed access to the parent tag which wasn't in the stubs yet.