Fill date to NewsItemInfo.timestamp when available
Parse the dates from h2 headers in the news page and store them to a
timestamp field in NewsItemInfo objects.

Some news item links don't have a date in the news page, so set the
timestamp field to None for those.

Alse extend the bs4.element.Tag type stubs with a couple new methods,
since one of them was needed in the implementation.