Fix parsing of news item title with a "New" label
Sometimes there is a span with text "New" before the subject text in the
title element.  This messed up the subject parsing, since the "New" text
and an extra newline character were prepended to the subject.

Fix this issue by removing such labels from the subject, and also mark
the NewsItemInfo as is_unread (a new attribute) when such "New" label is
found.